+++ src/components/shared/Footer.astro
@@ -47,7 +47,7 @@ import Logo from "../ui/Logo.astro"
 		class="container mx-auto border-t border-t-secondary-900 py-2 text-center text-secondary-400"
 	>
 		<Typography type="p" class="small"
-			>© Clínica Smile Works. Todos los derechos reservados.</Typography
+			>© Clínica Smile Works. <br class="block sm:hidden" /> Todos los derechos reservados.</Typography
 		>
 	</div>
 </footer>
